+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.0.2] - 2019-11-04
+### Added
+- Winston logging with a verbose check
+- check to see if img exists
+- Can now read img file from json config
+
+### Fixed
+- error reporting image size equals zero is nog caught
+- cleaned up dist folder
+
 ## [0.0.1] - 2019-11-01
 ### Added
 - Initial setup and release